refactor : validationService로 유효성 확인 제거 (#125)
