--- breaditnow-customer-api/src/main/java/com/breaditnow/customer/product/application/port/LoadProductPort.java
@@ -1,9 +1,0 @@
-package com.breaditnow.customer.product.application.port;
-
-import com.breaditnow.customer.product.domain.Product;
-
-import java.util.Optional;
-
-public interface LoadProductPort {
-    Optional<Product> loadProduct(Long productId);
-}
