+++ breaditnow-customer-api/src/main/java/com/breaditnow/customer/product/domain/port/LoadProductPort.java
@@ -1,0 +1,10 @@
+package com.breaditnow.customer.product.domain.port;
+
+import com.breaditnow.customer.product.domain.Product;
+
+import java.util.Optional;
+
+public interface LoadProductPort {
+    Optional<Product> findProduct(Long productId);
+    Product getProduct(Long productId);
+}
